+++ app/src/main/java/com/example/bodybalance/videoplayer/domain/usecase/ExistsPlaylistVideoByIdUseCase.kt
@@ -2,5 +2,5 @@ package com.example.bodybalance.videoplayer.domain.usecase
 
 interface ExistsPlaylistVideoByIdUseCase {
 
-    suspend operator fun invoke(id: Double): Boolean
+    suspend operator fun invoke(id: Int): Boolean
 }
